Always create a new simulator when allocating
I was finding that the same UDID/simulator was being used over and over and eventually traced it back to this re-use here.
This changes it so that each invocation of allocate will create a new simulator. Ideally we'd have some way to choose whether to reuse or not from above but I'm not sure how best to structure that.