Change shebang to /bin/bash
The shebang used to specify /bin/sh, but the script itself uses non
standard features. It probably works most of the time because `sh` on a
lot of systems is just bash. E.g. on my macbook:

```
~ $ sh --version
GNU bash ...
~ $ bash --version
GNU bash ...
```

Instead of trying to make it a true `sh` script, let's just use
/bin/bash.