Make fbsimctl depend on FBSimulatorControlKit
Summary:The rationale for having duplicate sources was that we wouldn't end up with the swift shims duplicate link messages.
However, this is kind of pointless now given that we had this problem when Swifter was added to the targets.
Instead it's much cleaner to depend on a Single Framework. The only remaining problem is ensuring that the shims aren't statically linked.

Reviewed By: marekcirkos

Differential Revision: D3040905

fb-gh-sync-id: 697af4f101e2a4db3707b3e066f09800f7c59bf5
shipit-source-id: 697af4f101e2a4db3707b3e066f09800f7c59bf5