Fix Timescale Conversion
Summary: FBFramebufferVideo allows for videos to be in a ready state by keeping the last frame around, even when the video hasn't started yet. When the video does start, it is pushed at the current time. The problem was that we weren't updating the frame with the current time. This meant that depending on how long ago the last frame was, the timescale can go negative and screw the video up. This makes sense as the time that the frame was recieved from the framebuffer is *before* the start of the video. As a result it will have a negative time relative to the start of the video.

Reviewed By: nqmtuan

Differential Revision: D3065578

fb-gh-sync-id: 47676015f3d639bd561112c1e3d9744f755eef3d
shipit-source-id: 47676015f3d639bd561112c1e3d9744f755eef3d