Add 'Tap' Action
Summary: Adds a 'Tap' action that will use the `SimulatorBridge` to tap an element in the same manner that `Simulator.app` does. There are success/failability semantics that mean that if an element isn't located at a given position, it will fail. Moves all bridge interactions to their own interaction category so that the failure semantics of a missing bridge exist.

Reviewed By: marekcirkos

Differential Revision: D3108835

fb-gh-sync-id: 9fd11dfe8b9b465cd8459eb0ef4a37b39bd0a621
fbshipit-source-id: 9fd11dfe8b9b465cd8459eb0ef4a37b39bd0a621