Connect FBSimulatorBridge for Simulator.app launches
Summary: We can't connect the Framebuffer to an application that has been launched by Simulator.app, but it is possible to connect the bridge. This consolidates a lot of the startup path and means that Simulator.app launches can benefit from the SimulatorBridge interactions. This will also enable connecting a Bridge to a running simulator.

Reviewed By: marekcirkos

Differential Revision: D3116151

fb-gh-sync-id: 7da224d7dc39bb25259029941066f21d47dec38c
fbshipit-source-id: 7da224d7dc39bb25259029941066f21d47dec38c