Don't load the XCTestBootstrap private Frameworks from FBSimulatorControl
Summary: `XCTestBootstrap` has to load a whole bunch of Private Frameworks. Instead of requiring every time that `FBSimulatorControl` is used the Frameworks are loaded, we can load the frameworks when the `XCTestBootstrap` interaction is used. This will save on Framework loading time for invocations of `FBSimulatorControl` that don't use the `XCTestBootstrap` methods.

Reviewed By: marekcirkos

Differential Revision: D3139023

fb-gh-sync-id: e91fe2ce7b4d03a49441b42659a7fd7a39613dc3
fbshipit-source-id: e91fe2ce7b4d03a49441b42659a7fd7a39613dc3