Use Unisgned Integer (unsigned long long) for enumerations
Summary: `FBSimulatorState` puports to mirror the values of `CoreSimulator` but this isn't completely true as there is a difference in signedness. The dissasembly for CoreSimulator shows the `Invalid` state as being anything greater than `0x4` rather than `-1`. This means that Index Sets can be used as a more efficient way of storing collections of enumerated types, since the states can't be stored in a bitmask.

Reviewed By: marekcirkos

Differential Revision: D3184146

fb-gh-sync-id: 1ffe744611da91cb224b5c9b4cec6d8b9acfd408
fbshipit-source-id: 1ffe744611da91cb224b5c9b4cec6d8b9acfd408