Integration Tests for Running fbsimctl commands.
Summary: The Parsers for commands are tested themselves, but running multiple `fbsimctl` commands across multiple device sets on the same host is an effective way to get a high level of concurrency on a host. However, it's important to verify that the Simulators in separate device sets can operate independently of each other without colliding into each other. Since the device sets aren't `chroot`ed it's still possible for one command invocation to mess with another's simulators. With that in mind, I've added integration tests to verify that this doesn't happen

Reviewed By: migchez

Differential Revision: D3191047

fb-gh-sync-id: 811840cb836523a6089bfb116005bc422f389703
fbshipit-source-id: 811840cb836523a6089bfb116005bc422f389703