Use Options for enabling stdout/sterr redirection for Agents
Summary:Currently File Handles are provided to launched agents on the Simulator. This means that the containing process has to manage the lifecycle of the File Handles, when using files directly will:
1) Allow stdout/stderr logging to work for Applications. Only file paths are supported for Applications
2) Not make stdout/stderr logging special snowflakes when `FBSimulatorDiagnostics` is a good place to access the logs.

This also removes the file handle passing on the launch of Applications, since this isn't supported and doesn't (currently) work anyway.

Reviewed By: migchez

Differential Revision: D3195989

fb-gh-sync-id: c27402ed8202f6c7d01703c19bce3e1150f73244
fbshipit-source-id: c27402ed8202f6c7d01703c19bce3e1150f73244