Simplify Input Buffering by consolidating around CommandBuffer
Summary: The fact that Socket and File Handle relays need to know about linebuffering is leaky, it's perfectly possible to construct a valid command over multiple lines. These sources of input data should just shove the bytes into a blackhole and get results back. This diff trims down the amount of knowledege that sources of input data need to know about how to dispatch the data.

Reviewed By: marekcirkos

Differential Revision: D3218663

fb-gh-sync-id: 3ee9fa67d95167a4294dabeef47dd1e5ca9c7b02
fbshipit-source-id: 3ee9fa67d95167a4294dabeef47dd1e5ca9c7b02