Add FBTestManagerResultSummary
Summary: Instead of exposing a big old load of arguments when reporting some results, we can take some of the arguments and constuct an object that contains all of the values. This model can later be expanded to allow a query of all the finished results, so that implementors don't have to accumilate state.

Reviewed By: marekcirkos

Differential Revision: D3275487

fb-gh-sync-id: f36a52e7966f5f2b6a225cd9739c2a54a0c46cbd
fbshipit-source-id: f36a52e7966f5f2b6a225cd9739c2a54a0c46cbd