Add FBXCTestManagerLoggingForwarder
Summary: In an effort to decompose `FBTestManagerAPIMediator`, one of the first things to break out is the forwarding of the methods called in the `XCTestManager_IDEInterface` protocol. This will also make it easier to break cycles by placing an intermediate object that is strongly referenced by the Mediator, and has a weak backreference to the mediator.

Reviewed By: nqmtuan

Differential Revision: D3281070

fbshipit-source-id: dc5ae6329da49a6ed2bad93ee93bb94ceb0b631a