Correlate Simulator.app to UDID of Simulator not launched by FBSimulatorControl
Summary:
The `NSUserDefaults` for the iPhone Simulator.app launched by Xcode or by some way other than through `FBSimulatorControl` are stored in a Library location.
Since these methods of launching a Simulator imply that arguments or environment markers are not passed down to the launched process, we can use this effectively global
value in the defaults to determine the UDID of a Simulator.app, since there will only ever be one instance of Simulator.app applications of this kind.

This means that it is now possible to use `fbsimctl` commands such as `shutdown` and `erase` that should also teardown the `Simulator.app` if one exists.
The previous scenario was that the `launchd_sim` would dissapear as expected, but the `Simulator.app` would still be present with an un-interactable UI.
This is because the Framebuffer goes missing behind the back of the `Simulator.app` without cascading to killing of the Mac Application

Reviewed By: asm89

Differential Revision: D3292674

fbshipit-source-id: 652cf9683aa5b0c1606ae4c9d2d08e6beae80e3f