Add FBTestDaemonConnection
Summary: Breaks out the connection to the Test Daemon to a separate class. This splits out a bunch of the responsibility for establishing this connection and makes the connection easier to understand in success/failure terms. It will also allow refining of the internals of the connection being established since it should be easier to predict what is happening. In addition, the new class only needs to us it's slice of delegate methods as it forwards delegate messages back to the mediator/logger.

Reviewed By: nqmtuan

Differential Revision: D3292244

fbshipit-source-id: 7aaf213cff8dc05c4eda899dd5b08d5220721ce9