Add FBTestBundleConnection
Summary: In the same manner as `FBTestDaemonConnection`, `FBTestBundleConnection` manages the connection to the Test Bundle and therefore therefore the lifecycle of the 'Test Plan'. The coordination of these two classes can be performed by the caller. As the `FBTestBundleConnection` is the delegate of the `DTX` classes, it can view and then forward relevant messages to update it's own internal state. In the future this will allow the Test Plan to be started *after* the daemon is connected and controlled separately.

Reviewed By: nqmtuan

Differential Revision: D3297779

fbshipit-source-id: 1a07dd662fb15dd4d628f80b98b3ae5672c55857