Separate the Connection of Daemon and Execution of the Test Plan
Summary:
Currently, the connection to the testbundle/daemon are entangled with the starting of the Test Plan. This means that the start of the Test Plan cannot be deferred and that the caller has to spin a run loop in order to recieve the test run events. For many usages it makes more sense to call one method and wait for all the results to come back.

The current implementation doesn't change the semantics of the test lifecycle; events are delivered asynchronously when the test plan starts. The difference with before is that the connection being established is completely synchronous and the starting of the Test Plan is no longer an effect of connecting the daemon. The starting of the test plan is now an explicit call that is only valid when the state of the daemon is ready.

Reviewed By: nqmtuan

Differential Revision: D3304638

fbshipit-source-id: c8bb2305883b3e8617ac867b7d73e003676f2099