Add FBApplicationCommands
Summary:
In order to support binding with different implementations of interacting with iOS at runtime it makes sense to split up the `FBDeviceOperator` into sets of protocols that defines a simple API for interacting with iOS. This also moves the knowledge for this into `FBControlCore` so that `FBSimulatorControl` and `XCTestBootstrap` can depend on the same Framework.

The `Interaction` model really only serves to make the API more chainable, but is a gigantic set of categories with implementations of a bunch of different ways that a Simulator can be manipulated. Moving these into individual `Commands` whilst wrapping them with a chainable API will simplify the implementation of the interactions greatly.

Reviewed By: mmmulani

Differential Revision: D3305192

fbshipit-source-id: 3d2f6dc2ff0e0eafa47be179dd7575dde12cd431