Add MobileDeviceKit to linkable Private Frameworks
Summary: We need to be able to link Private Frameworks of 'Apple Configuration', which is separate from Xcode's Developer Directory. The values passed into the `FBWeakFramework` resolving methods are now provided in the constructor, so that callers can ignore passing in in fallback paths. This means that Frameworks across different base paths can be mixed and matched.

Reviewed By: asm89

Differential Revision: D3310154

fbshipit-source-id: 51b673d563d971cca5dd6fa6d726a24222a5ce89