Add FBAMDevice
Summary:
`MobileDevice` is a low-level Framework primarily with `CoreFoundation` types. It exists at the `/System/Library` level and is what `MobileDeviceKit` is driven by. It's also very fast when it comes to finding all devices so is worthy of being wrapped in `FBDeviceControl`. I suspect that it will also remain very stable across Mac OS X Versions.

We may be able to link less of the Xcode Frameworks in order for to perform simple tasks such as listing attached devices/udids and installing Applications. I suspect that we will need `DVT` for bootstrapping tests with `XCTestBoostrap`. A

It may be just as easy to use as `MobileDeviceKit`, which is the 'Apple Configurator' App's Framework that wraps this in an objective-c API. If we can avoid depending on `MobileDeviceKit` we will not need hosts running `FBDeviceControl` to depend on this.

Reviewed By: marekcirkos

Differential Revision: D3321927

fbshipit-source-id: d2892de5777e4e7f2717c27fa43a18b1729c4923