Use Message forwarding for Command Implementations
Summary:
Message forwarding is great for managing separate Command class implementations and dispatching to the appropriate one when the method is called on the `FBSimulator` instance.

This will also help out when we have common protocols for defining the availability of particular properties and functionality that is shared between `FBDevice` and `FBSimulator`.

Reviewed By: marekcirkos

Differential Revision: D3321930

fbshipit-source-id: 9533f74531ffb78433c3c6cc98f788693f42b743