Add FBiOSTarget
Summary: This can become the common protocol that Devices & Simulators provide. The Target can then expose a set of Commands, which will allow callers of the API to use the methods for Devices & Simulators. We can add to this Protocol over time as more common behaviours for Devices & Simulators become available.

Reviewed By: marekcirkos

Differential Revision: D3322073

fbshipit-source-id: fe308861975324f4ba750abdce9330ad63697d0d