Allow overriding of Locale/Language on App Launch
Summary:
It's possible to pass `NSUserDefaults` on the Command Line instead of setting globally for the whole Simulator.

Using a model to represent ways of setting the arguments via a defaults plist and other arguments. If we decide to later, we may be able to add additional configurable overrides for keyboards etc.

Reviewed By: marekcirkos

Differential Revision: D3328191

fbshipit-source-id: 37688698880aa33210f26f8e9581d1f90228dd00