Move Shared Configuration to it's own Config
Summary: Instead of the copypasta of the `xcconfig` files, it makes more sense to move these definitions to a shared configuration, which is then `#include`d from the other configurations. The configuration that is currently in the `FBSimulatorControl.xcodeproj/project.pbxproject` can then be extracted out to the per-target files. This can aid migration to GitHub buck builds and possibly a public workspace instead of a massive project.

Reviewed By: marekcirkos

Differential Revision: D3365232

fbshipit-source-id: b8234da1abff44c8525d9298599d632dba928fa2