Bypass Loading of User Plugins
Summary: User plugins, which are plugins at `~/Library/Application Support/Developer/Shared/Xcode/Plug-ins`, can cause issues if linked into the same image as a process that already references the same symbols. In order to save time and prevent runtime linker issues it makes sense to bypass the loading of these symbols. The simplest way of doing this is swizzling the bundle loader method and log/no-op in the event we have one of these bundles.

Reviewed By: mmmulani

Differential Revision: D3392950

fbshipit-source-id: cf173c8bd548d2b26d8cbc4dbb27d226582bf036