Add FBiOSTargetDescription
Summary: Moves the specifier for a Device/Simulator's description to a concrete, immutable class. This means that the same description can be used to render the formatting of both devices and Simulators. Additionally this class now becomes a first-class citizen and `FBSimulatorControlKit` can use it directly, just like the promotion of `FBSimulatorQuery`. The serialization also means that the description can be serialized over JSON to provide the consumer of the HTTP server with a slice of information about the Simulator on the remote end.

Reviewed By: marekcirkos

Differential Revision: D3404353

fbshipit-source-id: 81be85020cb834874e1bd262501759e374c467c5