Set INSTALL_PATH in Frameworks
Summary:
I didn't know this before today, but it looks like the `INSTALL_PATH` contributes to the way that dependent Frameworks add load commands in their Mach-O header.

In essence:
1) `FBSimulatorControl` had an install path of the default `/System/Library/Frameworks`, this is derived from `INSTALL_PATH` and results in the `LC_ID_DYLIB` header getting set to this value (P56410820)
2) When a dependent Framework links and bundles `FBSimulatorControl`, it references the location from this Mach-O header. In this case it meant that `FBSimulatorControlKit` was referencing `FBSimulatorControl` as if it were installed at the system level.
3) When `fbsimctl` goes through the dynamic linking process, the dylib load references a Framework that doesn't exist at the specified path.

This is documented at: http://bob.ippoli.to/archives/2005/01/05/versioned-frameworks-considered-harmful/

'When you create an Mach-O object file (executable, dylib, bundle, etc.) that depends on another, it will generate a load command (LC_LOAD_DYLIB) referencing the other dylib. The data associated with the LC_LOAD_DYLIB is exactly the same data that was provided by the linked-to MH_DYLIB in its LC_ID_DYLIB load command'

This code change ensures that the `LC_ID_DYLIB` header in `FBSimulatorControl` is relative to the `rpath` so that it can be loaded correctly by `FBSimulatorControlKit`. `FBSimulatorControl` bundles `FBSimulatorControl` in it's Frameworks directory. The `FBSimulatorControlKit.framework/Frameworks` directory is in the `rpath` of `FBSimulatorControlKit`'s executable so it will now load just fine.

Reviewed By: migchez

Differential Revision: D3404816

fbshipit-source-id: 8b187dc40897b06bece2aa73f7f25495fd821e21