Extract SimulatorSubject to iOSTargetSubject
Summary: Making `SimulatorSubject` more generic and intead defined in terms of an `iOSTargetSubject` will mean that `FBDevice` can be shoved into this Subject as well as `FBSimulator`. I was considering making this generic with a type constraint to `FBiOSTarget`, however this doesn't seem crucial right now as the `Subject` itself is passed to the event reporting mechansim. As such, the fact that the target is a 'Simulator' isn't important, Reporters only care that there are a bunch of properties to extract information about the state of the iOS Target.

Reviewed By: nqmtuan

Differential Revision: D3410952

fbshipit-source-id: a5ead47013d8700706a3eead83d094c4727b611f