Extract SimulatorReporter to iOSReporter
Summary:
Extracting `SimulatorReporter` functionality to `iOSReporter` will allow us to create a device-specific reporter for reporting events to, without caring that the underlying iOS Target is a Device/Simulator. I have made this protocol generic with a type constraint since `SimulatorReporter` is used as both a way of reporting events and a container for the Simulator itself. This is a bit ambiguous but it allows the `SimulatorReporter` to be passed around to `Runner` classes that do the work of performing some `Action` and reporting on it.

Given that performing the Action itself is not so abstact that it closes over the iOS Target to which the Action applies, I've not changed this behaviour yet.

Reviewed By: nqmtuan

Differential Revision: D3410958

fbshipit-source-id: 3d347d0d4b7c8577247195267b6d670b275f15dc