Extract SimulatorRunner to iOSTargetRunner
Summary:
In further preparation for Device support, the `SimulatorRunner` can be extracted to a more generic `iOSTargetRunner`. This will allow for Devices to use the same `iOSTargetRunner` as Simulators, with all the associated reporting the start/end/error of any given action. The `iOSTargetRunner` also allows actions to be defined in terms of another Runner implementation, instead of a function that returns a `CommandResult`.

A `SimulatorRunner` is essentially 'Given a Simulator some Event Names and a Closure, report on the start/success/failure of the closure with the event names'

Reviewed By: nqmtuan

Differential Revision: D3410960

fbshipit-source-id: a4480717b3345920a832315bb4bb2f4e84c8ba86