Extract Command to a Struct
Summary: This is partly to appease Swift 3 which is currently segfaulting when compiling the  `Command` enum. It makes sense to have this in a struct anyway, since the associated types in an enum become unwieldy as they are unnamed. Pulling `Command` out to a struct means that the `CLI` can be non-recursive and `Help` can reference a `Command`.

Reviewed By: marekcirkos

Differential Revision: D3496858

fbshipit-source-id: 451d3d9d1ff240844c9f7b42375601569acc41c6