Add FBiOSRunnerContext
Summary: There's a lot of shared values that have to get passed down through each of the runners. Having these as arguments to the constructor is reasonably terse, however there's still a considerable amount of context that can be encapsulated in one structure. Passing a Generic Context down through each of the runners makes the whole binding more terse and will allow adding a `FBDeviceSet` to the context considerably easier, since the `FBDeviceSet` value won't have to be added to each of the runners but in one place alone.

Reviewed By: marekcirkos

Differential Revision: D3425033

fbshipit-source-id: 7eae42927e722fcffd124e409e090f36e8aa479c