Remove redundant arguments from BaseCommandRunner
Summary: The `BaseCommandRunner` acts as a root runner for all `Commands`, in that it gets all of the dependencies needed to actually run a `Command`. This is distinct from `CommandRunner` which requires that the `FBSimulatorControl` Framework is allready loaded, which makes it a better fit for the `Server` case. This removes the need to have optionals at the `struct` level and pass these as non-optional values to another `Runner`.

Reviewed By: marekcirkos

Differential Revision: D3502979

fbshipit-source-id: 072b36d604da1f947bb435290ed5eeec71a700a8