Fix Test Failure if Device is Connected
Summary:
I probably need to re-think how we can select devices/simulators/all (perhaps remove it from being a switch). But this will at least fix the integration test, that assumes that there will be no results returned.
I think the situation that makes the most sense is to avoid loading `FBDeviceControl` when a custom Simulator set is used.

Reviewed By: marekcirkos

Differential Revision: D3514684

fbshipit-source-id: 83b33f80ffbe4f7e81a1f8fd6266b969e971adc4