Add CreationConfiguration
Summary: Adds an element of indirection to the construction of a Configuration for creating a Simulator. This means that the loading of the `defaultConfiguration` can be deferred until after the `CLI` is Parsed. This means that we can defer the loading of `CoreSimulator` until we actually need it. Additionally, we get the benefits of the `Accumilator` protocol, which means that the ordering of the arguments after the `create` action is not important.

Reviewed By: marekcirkos

Differential Revision: D3514686

fbshipit-source-id: f4443c9458c2d9bc8bbd20043d7a89b85f38cf05