Don't load FBDeviceControl if a custom Device Set is used.
Summary: A custom `--set` passed on the CLI means that the caller is expecting to be able to manage isolated Simulators, in which case loading a `FBDeviceSet` makes no sense. Since loading `MobileDevice` loads a bunch of frameworks (including blocking on `usbmuxd`) we should avoid loading these Frameworks entirely. nil coalescing can be used to avoid drastically altering the fetching process. Additionally, the same loading mechanism is used to thread through the logger as with `FBSimulatorControlFrameworkLoader`

Reviewed By: marekcirkos

Differential Revision: D3514689

fbshipit-source-id: 15d1c58b4390d5e30e882ed913ed0959e0b98ad4