Pass FBDevice to FBiOSDeviceOperator instance.
Summary:
This is a starting point to extracting out some of the Device Operator commands to specific protocols that `FBiOSDeviceOperator` can then forward to. `FBiOSDeviceOperator` currently exists as two things:
1) A implementation of a bunch of important commands for making `FBDeviceControl` actually useful
2) An implementation of the `FBDeviceOperator` Protocol to make do some of the heavy lifting for `XCTestBoostrap`.

I want to separate 1 from 2, so `fbsimctl` can use 2 directly without having to care about 1. The accessor for `dvtDevice` on `FBDevice` will automagically instantiate the DVT class where appropriate. It may also be possible to avoid using `dvtDevice` for a subset of commands and bypass the use of Xcode Private Frameworks for these purposes

Reviewed By: marekcirkos

Differential Revision: D3516166

fbshipit-source-id: cca46aa430985a4f67c750c1bc6165713b0892e8