Use Framework Versioning to prevent Runtime Duplicate Symbols
Summary:
`FBControlCore` has some provisional Framework Version build settings enabled and is the only of the 4 Objective-C Frameworks to do so. `XCTestBootstrap` does not have versioning, but is embedded in both `FBSimulatorControl` and `FBDeviceControl` just like `FBControlCore` is. This poses a problem when `FBSimulatorControl` and `FBDeviceControl` are both loaded into the same process:

1) `fbsimctl` loads `FBSimulatorControl`, `dyld` loads the Frameworks in `FBSimulatorControl.framework/Frameworks` of `FBControlCore` & `XCTestBootstrap`.
2) `fbsimctl` loads `FBDeviceControl`, `dyld` loads the the Frameworks in `FBDeviceControl.framework/Frameworks`.
3) `FBControlCore` is not loaded as it has the same Framework version as was loaded in #1
4) `XCTestBootstrap` is loaded *again* since it has no versioning information
5) `dyld` complains about duplicate symbols at runtime.

Adding Versioning to `XCTestBootstrap` has the effect that `dyld` will be smart enough to not complain about duplicate symbols in `XCTestBootstrap`. This means that we can keep `FBSimulatorControl` as being completly (and independently) redistributable as a Framework as it will continue to bundle its dependencies, but also prevent runtime duplicate symbols.

Reviewed By: marekcirkos

Differential Revision: D3533943

fbshipit-source-id: bf9bc7562a1a4a6275d6d6089bf2da4825c9637e