FBSimulatorBridge -> FBSimulatorConnection
Summary: This reduces the ambiguity with the actual `SimulatorBridge`, which is a subset of the actual services that can be connected (and some independently). It looks like connecting the `SimulatorBridge` can be somewhat expensive, so in many circumstances we may be able to bypass it and lazlily connect the bridge.

Reviewed By: marekcirkos

Differential Revision: D3522037

fbshipit-source-id: bebe3cbdbcb484980339949272be998acd5d2ace