Bridge Connection can be lazily established.
Summary:
The `FBSimulatorBridge` is a remote object connection that does not need to be established when the Simulator is booted. For this reason the `FBSimulatorBridge` component does not have to be instantiated on boot, though this is currently a requirement of direct launching. This can increase the performance of booting if a bridge connection is not required. This change splits the instantiated of composed connection properties into two stages:
1) The Instantiation stage, which takes arguments that are created/fetched before/after the booting of a simulator.
2) The Bridge Connection stage, which can attach a bridge regardless of the arguments in the Instantiaton phase.

Reviewed By: marekcirkos

Differential Revision: D3533994

fbshipit-source-id: 9ab761dfc9f3ed3e3c40c448221f7c2e87f9eea5