Improve Simulator Connection API
Summary:
Fixes https://github.com/facebook/FBSimulatorControl/issues/276. With the API changes towards `FBSimulatorConnection`, I missed the 'Connecting to an allready booted Simulator' case. This fixes this by allowing lazy instantiation of the `FBSimulatorConnection`.

Additionally, this extracts out the intricacies of actually disconnecting `FBSimulatorConnection` to the same source category as establishing the connection.

Reviewed By: marekcirkos

Differential Revision: D3543263

fbshipit-source-id: 7d5c408def4d4bbc0ada7eff66646760eb409616