Prevent deadlock in Simulator Teardown
Summary:
There's a latent condition inside the Simulator teardown path that means that the `FBTestDaemonConnection` is dellocated after a Simulator has shutdown. This is something that needs to be avoided as the process can deadlock in the `-[FBTestDaemonConnection .cxx_destruct]` when the `deamonProxy` property is deallocated. It appears that the issue is reproducable before the `FBSimulatorConnection` changes, but is far easier to reproduce after these changes.

The erroneous behaviour is that the TestManagers are only tidied up when the Simulator is launched via the `Simulator.app`. This is something that needs to happen for all TestManagers that are connected to any terminating Simulator.

Reviewed By: marekcirkos

Differential Revision: D3549926

fbshipit-source-id: 2f4d69ad1ecd9353fd02f926a197d2025d3cb059