Add CLI Tests for fbsimctl
Summary:
There are a number of Unit & Integration tests for `fbsimctl` via the `FBSimulatorControlKit` target. These tests however, don't test the full lifecycle of the process from start to finish. `FBSimulatorControlKitTests` all execute within one process lifecycle and do not test the main entry-point of the Application.

This diff adds the basics of a Python3 based e2e-testing suite for testing the `fbsimctl` executable through multiple invocations of the `fbsimctl` executable. It uses the `--json` argument and wraps these invocations in python objects and conveniences in order to make driving the tests easier. There is a base `unittest` class that adds a number of assertions that use these utilities. It should be easy to take this base and extend it to test many of the actions and automation scenarios for `fbsimctl`. I may also look at adding some nice logging to make it possible to understand how `fbsimctl` is being driven, which may help developers who wish to integrate `fbsimctl` into their own projects.

I will follow this up with extensions to the build script and travis manifest, so that we can run this on travis ci. The test script currently assumes that `fbsimctl` is on the path, but we will be able to change this to point at a build that we create in CI and pass this through, so we can test regressions on the cli iteself

Reviewed By: migchez

Differential Revision: D3569094

fbshipit-source-id: d7680e124dbfe1595498095110f5b7c2ef7c3799