Add 'Missing Default' Simulator Creation
Summary: New installs of `Xcode` will have a default set of Simulators created that represents all the Simulator Devices, for each available Runtime. This array can be constructed by enumerating along each of the runtime/devicetype axes. I've added an API for getting all these configurations, as well as only the ones that are missing from the current set.

Reviewed By: marekcirkos

Differential Revision: D3612809

fbshipit-source-id: 63e67682d061812fcf071d3052078437a73b3405