Defer reporting 'Listen' has started until relay starts.
Summary:
Consider the following.
1) The user has a `listen --http 8090` action.
2) 'listen' 'started' is reported.
3) The consumer of `fbsimctl` sends a `SIGTERM`
4) The Signal handler is installed
5) The event loop starts and the process never appears to respond to #3

By defering the reporting of the 'listen' 'started' event, we can be certain that the signal handlers have been installed. I found this race when working on the cli-e2e tests, which can rapidly close the process after they determine 'listen' 'started', in this case the calling process would hang since the `SIGTERM` would race with the startup of the HTTP server.

Reviewed By: marekcirkos

Differential Revision: D3635123

fbshipit-source-id: c31aa7b9ce5d185035d9640cdda89a06783124e3