Fix launchctl wrapper process launching.
Summary: The wrong `launchctl` executable was being launched on the Simulator for iPhone/iPad devices. Additionally the reading of the file handle was excessively complicated. The timeout provided by the wrapper will ensure that the pipe is not blocked, since the write end of the file handle is closed before reading from the read end of the pipe.

Reviewed By: marekcirkos

Differential Revision: D3674684

fbshipit-source-id: 822e9a47a585a2d0e2c005ead2249dcea2cb5bea