Extract and annotate FBBatchLogSearch
Summary:
`FBLogSearch` provides a high-level and performant API for searching through a bunch of test. I wish to use this for parsing out a single line from `launchctl`. Instead of re-inventing the wheel when it comes to extracting matches, I'd rather just use `FBLogSearch` to extract a single line, then parse the output of the single line.

This diff extracts `FBBatchLogSearch` to a separate file, so that `FBLogSearch` can be adapted into a class-cluster that can deal with non-diagnostic input.

Reviewed By: marekcirkos

Differential Revision: D3674708

fbshipit-source-id: 783e64983b42a9becc6d7bfe34c2a87c8d429b09