Updates for Xcode 8 Beta 5
Summary: In Beta 5, the Frambuffer is provided by a call to `setIOSurface:` on the client of the `SimDeviceFramebufferService`. This can be seen in `Simulator.app`'s `-[SimulatorViewController setIOSurface:]` call. This is also suggested by the method name `-[SimDeviceFramebufferService _ON_EXECUTION_QUEUE_sendSetIOSurfaceToClients:]` method. This is shown in the backtrace, when `setIOSurface:` is called. Currently, this diff just ensures that we can recieve the `IOSurface` and logs when we do. Later on, we'll have to use this to extract frames or render onto a `CIImage` like `Simulator.app` does.

Reviewed By: marekcirkos

Differential Revision: D3696257

fbshipit-source-id: 42f479ab009fcc81755fdcb550b74667fdc3d121