Improve Startup Performance by fetching Processes at once.
Summary:
The slowest part of the launch path for `FBSimulatorControl` for a Simulator Set with a non-trivial number of Simulators is the fetching of the `launchd_sim` and `Simulator.app` processes that are associated with a Simulator. This is currently done implicitly, on the instantiation of a Simulator. By fetching all these processes up-front and then passing them down on the instantiaion of a `FBSimulator` instance, we can turn a linear-time operation to a constant-time one.

From my simple tests in the shell, it has ~2X improvement of the time taken in `hg list`.

Reviewed By: marekcirkos

Differential Revision: D3689890

fbshipit-source-id: 958171df078bc7aa876f407632100b16228e45b4