Fix bug in Framework Loader
Summary: `XCTestBootstrapFrameworkLoader` was using the wrong method name for the subclassing of `FBControlCoreFrameworkLoader` this was causing `launch_xctest` to fail, since `DTXConnectionServices` wasn't loaded. This wasn't caught on CI since there is currently no e2e test for `fbsimctl` to `launch_xctest`. When running in a test context, these frameworks are loaded by XCTest inside the same process

Reviewed By: nqmtuan

Differential Revision: D3709107

fbshipit-source-id: 7c3f309e492b4d180b742abcfa7d590c3c8a379c