Don't 'persist' direct launched Simulators
Summary: This can cause all manner of havok as it breaks the current expectation that when a container process goes down, it can take a Simulator with it. This may be something that we can take advantage of in the future if it means that we can have a container process as well as other processes using the same Simulator, but for now we should keep the semantics constistent. This being passed in `simctl` makes sense so that Simulators are kept alive over multiple invocations of `simctl` without a container `Simulator.app`

Reviewed By: nqmtuan

Differential Revision: D3703016

fbshipit-source-id: 0bf0f72d2f45717edafb3a0f2843b1c346ef88a8