Add FBFramebufferIOSurfaceFrameGenerator & FBFramebufferBackingStoreFrameGenerator
Summary:
I've now extracted each of the two different flows to separate Frame Generators. This means that consumers of Frames can use exactly the same API as before, but now we're using `IOSurface` and `CIImage for scaling and rendering.

I've split each of the kinds of Framebuffers into a class cluster, so that we'll get a much clearer picture of what the expected contract is with the `SimDeviceFrameBufferService`. This also splits out the abstract lifecycle from the individual concrete actions that need to be taken inside the `FBFramebuffer`.

Reviewed By: nqmtuan

Differential Revision: D3702825

fbshipit-source-id: 4d22808f9615578c146a1e896b49ce1858634f23